add model extended by phpcr document
refactoring code for CS

routes to model

replace name property by title

remove unused test document loading
