+++ Resources/config/doctrine-phpcr/SeoAwareContent.phpcr.xml
@@ -5,28 +5,11 @@
         https://github.com/doctrine/phpcr-odm/raw/master/doctrine-phpcr-odm-mapping.xsd"
         >
 
-    <document
-            name="Symfony\Cmf\Bundle\SeoBundle\Doctrine\Phpcr\SeoAwareContent"
-            referenceable="true"
-            translator="attribute"
-            >
-        <node name="node"/>
-        <locale name="locale" />
-        <nodename name="name"/>
-        <parent-document name="parent"/>
+    <document name="Symfony\Cmf\Bundle\SeoBundle\Doctrine\Phpcr\SeoAwareContent">
         <id name="id">
             <generator strategy="PARENT"/>
         </id>
-        <field name="title" type="string" translated="true"/>
-        <field name="body" type="string"  translated="true"/>
-        <field name="seoMetadata" type="string" assoc="" nullable="true" />
-        <referrers name="routes" referring-document="Symfony\Cmf\Bundle\RoutingBundle\Doctrine\Phpcr\Route" referenced-by="content">
-            <cascade><cascade-persist /></cascade>
-        </referrers>
-        <lifecycle-callbacks>
-            <lifecycle-callback type="preUpdate" method="preFlush"/>
-            <lifecycle-callback type="prePersist" method="preFlush"/>
-            <lifecycle-callback type="postLoad" method="postLoad"/>
-        </lifecycle-callbacks>
+        <node name="node"/>
+        <parent-document name="parent"/>
     </document>
 </doctrine-mapping>
